Remove the pre-release step
The -dev releases get in the way a lot because jitpack refuses to build
a prod release off the same commit it just built with a -dev tag.

This leads to people using -dev dependencies because nothing else is
available.

Since there is not much benefit to the -dev tags, I'm removing them for
now.