Add overrideUserId field to EventWithUserId
As part of this story we're enabling overriding a user id for
integrations that use a private client (these were always coming in as
the same user, which was the service account).

The original plan was to just use what they provide us in the
Boclips-User-Id header and set that in the userId field of events.

But these users aren't necessarily synced with our user database...
Meaning if we just blindly replace the id, we have no idea who that user
is.

Will store the id provided via header in overrideUserId field for now,
we can improve on this later on.

[#170026268]